Renamed hook useDesktopStatus to useIsDesktop.
