+++ src/scenes/Services/components/Flow/Flow.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import Switcher from './components/Switcher';
 import List from './components/List';
-import { useDesktopStatus } from '@hooks';
+import { useIsDesktop } from '@hooks';
 
 import styles from './Flow.module.scss';
 
@@ -33,7 +33,7 @@ const Flow = () => {
       message: `6) Halo Lab development process starts with understanding of your project and your needs, bottlenecks, and develop an appropriate estimation and project plan.`,
     },
   ];
-  const { isDesktop } = useDesktopStatus();
+  const { isDesktop } = useIsDesktop();
 
   return (
     <div className={styles.container}>
